# lib/api-zod/src/generated/types/authResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Talim Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { AuthResultRole } from './authResultRole';

export interface AuthResult {
  id: string;
  role: AuthResultRole;
  full_name: string;
  login: string;
  /** @nullable */
  class_name?: string | null;
  /** @nullable */
  class_id?: string | null;
  /** @nullable */
  telegram_id?: number | null;
  /** @nullable */
  token?: string | null;
}