# lib/api-zod/src/generated/types/dashboardStats.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Talim Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { ClassCount } from './classCount';

export interface DashboardStats {
  total_students: number;
  total_classes: number;
  total_staff: number;
  days_until_launch: number;
  students_by_class: ClassCount[];
}