# lib/api-zod/src/generated/types/myClassResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Talim Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { Student } from './student';

export interface MyClassResult {
  class_name: string;
  /** @nullable */
  class_id?: string | null;
  students: Student[];
}