# lib/api-zod/src/generated/types/schoolClass.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Talim Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface SchoolClass {
  id: string;
  name: string;
  /** @nullable */
  teacher_id?: string | null;
  /** @nullable */
  teacher_name?: string | null;
  student_count?: number;
  created_at: string;
}