# lib/api-zod/src/generated/types/staffInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Talim Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { StaffInputRole } from './staffInputRole';

export interface StaffInput {
  full_name: string;
  role: StaffInputRole;
  /** @nullable */
  class_id?: string | null;
}